requirements.txt
discogs_client
pandas
# psycopg2
numpy
seaborn
psycopg2-binary
python-dotenv
